split sql into distinct connectors
